fix: mistakes process pids

+++ src/data/binData.ts
@@ -51,9 +51,11 @@ class BinDataModel implements IBinData {
   killProcess(pid: string) {
     const childProcess = this.processes[pid];
 
-    log(
-      `Process ${pid} with pid:${this.processes[pid].pid} not dead, attempting to kill`
-    );
+    if (!childProcess) {
+      log(`Process with ${pid} can't be killed, already dead`);
+      return;
+    }
+
     if (os.platform() === "darwin") {
       if (childProcess.kill()) {
         console.log(`Process ${pid} killed`);
@@ -63,12 +65,8 @@ class BinDataModel implements IBinData {
     } else {
       const _pid = childProcess.pid;
       child_process.exec("taskkill /pid " + _pid + " /T /F");
-      if (!childProcess.pid) {
-        console.log(`Process ${pid} killed`);
-        return;
-      }
 
-      console.log(`Process ${pid} couldn't be killed`);
+      console.log(`Killing process ${pid} with pid${childProcess.pid}`);
     }
   }
 
@@ -84,57 +82,44 @@ class BinDataModel implements IBinData {
     options?: Record<string, unknown>
   ) {
     const fullCommand = path.join(this.binaryPath, command);
-
-    /**
-     * TO-DO: Needs to be tested
-     * For windows command should be something like:
-     * command => cmd.exe (entry point for command line, not the parameter command)
-     * args => ['c/', paramCommand, paramArgs]
-     *
-     * return child_process.spawn('cmd.exe', ['c/', command, ...args]);
-     *  var ffmpeg = spawn( 'cmd.exe', ['/c',  '"'+ffmpegpath+ '"', '-i', filelist[i], '-c:v', 'libx264', '-pix_fmt', 'yuv420p', '-c:a', 'aac', '-y', outfile]));
-     */
-
+    let process: child_process.ChildProcessWithoutNullStreams;
     if (os.platform() === "darwin") {
-      const process = child_process.spawn(fullCommand, args, options);
-
-      if (pid) {
-        this.processes[pid] = process;
-        log(`ChildProcess ${pid} with pid:${this.processes[pid].pid} created`);
-        process.on("exit", (code, signal) => {
-          if (signal === "SIGTERM") {
-            console.log(
-              `Child process ${pid} with pid:${process.pid} was killed with SIGTERM`
-            );
-          } else if (signal === "SIGINT") {
-            console.log(
-              `Child process ${pid} with pid:${process.pid} was killed with SIGINT`
-            );
-          } else {
-            console.log(
-              `Child process ${pid} with pid:${process.pid} exited with code ${code}`
-            );
-          }
-          delete this.processes[pid];
-        });
-      }
-
-      return process;
+      process = child_process.spawn(fullCommand, args, options);
     } else if (os.platform() === "win32") {
-      const process = child_process.spawn(
+      process = child_process.spawn(
         "cmd.exe",
-        ["/c", fullCommand, ...args],
-        options
+        ["/c", path.basename(fullCommand), ...args],
+        {
+          ...options,
+          shell: true,
+          cwd: path.dirname(fullCommand),
+        }
       );
-
-      if (pid) {
-        this.processes[pid] = process;
-      }
-
-      return process;
     } else {
       throw new Error("Platform not supported");
     }
+    if (pid) {
+      this.processes[pid] = process;
+      log(`ChildProcess ${pid} with pid:${this.processes[pid].pid} created`);
+      process.on("exit", (code, signal) => {
+        if (signal === "SIGTERM") {
+          console.log(
+            `Child process ${pid} with pid:${process.pid} was killed with SIGTERM`
+          );
+        } else if (signal === "SIGINT") {
+          console.log(
+            `Child process ${pid} with pid:${process.pid} was killed with SIGINT`
+          );
+        } else {
+          console.log(
+            `Child process ${pid} with pid:${process.pid} exited with code ${code}`
+          );
+        }
+        delete this.processes[pid];
+      });
+    }
+
+    return process;
   }
 }
